+++ src/modules/animation/animal/bird/anim/Idle.cpp
@@ -13,7 +13,7 @@ namespace animation {
 namespace animal {
 namespace bird {
 namespace idle {
-void update(float animTime, BirdSkeleton &skeleton, const BirdSkeletonAttribute &skeletonAttr) {
+void update(double animTime, BirdSkeleton &skeleton, const BirdSkeletonAttribute &skeletonAttr) {
 	const float sine = glm::sin(animTime);
 	const float cosine = glm::cos(animTime);
